Identifying and disabling pro features. The features need to be seperated out into the pro repo
Started the separation of the free and the pro features.
