+++ admin/src/css/4.regions/seocentral-redirects.css
@@ -343,4 +343,36 @@
     &-wrapper {
     }
   }
+
+  &-lite {
+
+    &-redirect {
+
+      &-wrapper {
+        @apply !relative;
+
+        .seo-central-redirect-wrapper, seo-central-redirection-wrapper {
+          @apply !pointer-events-none !cursor-default;
+        }
+        
+        .seo-central-redirection-wrapper {
+
+          .seo-central-delete-redirect, .delete, .seo-central-quickedit-redirect, .quickedit {
+            @apply !pointer-events-none !cursor-default;
+          }
+        }
+      }
+
+      &-overlay {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background-color: rgb(240, 240, 241);
+        opacity: 0.5;
+        z-index: 20;
+      }
+    }
+  }
 }
